fingerprint install

+++ services/Auth.js
@@ -12,8 +12,7 @@ import UserRepository from "../repositories/User.js";
 import { ACCESS_TOKEN_EXPIRATION } from "../constants.js";
 
 class AuthService {
-  static async signIn({ userName, password }) {
-    // fingerprint
+  static async signIn({ userName, password, fingerprint }) {
     const userData = await UserRepository.getUserData(userName);
     if (!userData) {
       throw new NotFound("Пользователь не найден");
@@ -30,8 +29,8 @@ class AuthService {
     await RefreshSessionsRepository.createRefreshSession({
       id: userData.id,
       refreshToken,
+      fingerprint,
     });
-    // fingerprint,
 
     return {
       accessToken,
@@ -40,8 +39,7 @@ class AuthService {
     };
   }
 
-  static async signUp({ userName, password, role }) {
-    // fingerprint
+  static async signUp({ userName, password, fingerprint, role }) {
     const userData = await UserRepository.getUserData(userName);
     if (userData) {
       throw new Conflict("Пользователь с таким именем уже существует");
@@ -58,7 +56,7 @@ class AuthService {
     await RefreshSessionsRepository.createRefreshSession({
       id,
       refreshToken,
-      // fingerprint,
+      fingerprint,
     });
     return {
       accessToken,
@@ -71,8 +69,7 @@ class AuthService {
     await RefreshSessionsRepository.deleteRefreshSession(refreshToken);
   }
 
-  static async refresh({ currentRefreshToken }) {
-    // fingerprint
+  static async refresh({ fingerprint, currentRefreshToken }) {
     if (!currentRefreshToken) {
       throw new Unauthorized();
     }
@@ -82,9 +79,9 @@ class AuthService {
     if (!refreshSession) {
       throw new Unauthorized();
     }
-    // if (refreshSession.finger_print !== fingerprint.hash) {
-    //   throw new Forbidden();
-    // }
+    if (refreshSession.finger_print !== fingerprint.hash) {
+      throw new Forbidden();
+    }
     await RefreshSessionsRepository.deleteRefreshSession(currentRefreshToken);
 
     let payload;
@@ -104,8 +101,8 @@ class AuthService {
     await RefreshSessionsRepository.createRefreshSession({
       id,
       refreshToken,
+      fingerprint,
     });
-    // fingerprint,
     return {
       accessToken,
       refreshToken,
